Remove $value and $ref from urls when matching permissions (#591)
* Add helpful comment; rethrow exception to preserve stack details

* Remove ${values} from incoming urls; refactor url formatting to common method

* Update tests to validate removing of ${value} segment from paths

* Remove Argument exception handling in permissions store

We want to capture all argument exceptions in the controller. 1) ArgumentNullException already gets captured in the controller. 2) ArgumentException needs to be captured as a server error because this is caused by a Regex related issue in the UriTemplateMatcher.

* Remove InvalidOperationException block

Because we only log the exception message and throw an error 500 - server error, we can combine this with the global exception block.

* Add helpful comment

Co-authored-by: Irvine Sunday <irochand@microsoft.com>