Replaces UriTemplateMatcher with OpenApiUrlTreeNode in OpenAPIService (#592)
* Add utils.cs for common solution-wide functionality

Helps with validating whether input is null or empty

* Bump lib.

To make use of the OpenAPIUrlTreeNode structure

* Replace UriTemplateMatcher with OpenApiUrlTreeNode

Also updates the controllers respectively to accommodate the change in method params.

* Update tests to account for method param update

Also updates a test to validate that paths are matched even with mismatched parameter names.

* Guard clause in expression

* Update extension method and its refs; add new test to validate

Update the UriTemplatePathFormat method to include a bool flag to either resolve fully qualified names or not. OpenApiService url resolving doesn't require namespace simplification.

* Whitespace fix

* Update OpenAPI yaml description

* Avoid recursively walking AdditionalProperty schemas

Because 'additionalProperties' are now being walked, we need a way to keep track of visited schemas to avoid endlessly creating and walking them in an infinite recursion.

* Get base uri from requests; add helpful comments

* Enrich tests to validate OpenApiUrlTreeNode; use Linq to get path operations

We're using Linq to get to the path operations instead of the path key -> the url because of the potentially different url value to the one defined in  the OpenAPIDoc. This throws a KeyNotFound exception

* Add new test path

* Ensure we simplify qualified namespaces when matching for permissions

* Another whitespace fix

* Updates XML documentation

* Update OpenAPIService/OpenApiService.cs

Co-authored-by: Vincent Biret <vibiret@microsoft.com>

* Update OpenAPIService/PowershellFormatter.cs

Co-authored-by: Vincent Biret <vibiret@microsoft.com>

* Remove async modifier from method

* Fix tests

Co-authored-by: Irvine Sunday <irochand@microsoft.com>
Co-authored-by: Vincent Biret <vibiret@microsoft.com>