Remove telemetry and exception throwing for every scopes info failure
Throwing exceptions every time the ScopesInformationDictionary is empty is not feasible. We end up getting a lot of availability alerts and exceptions which in turn trigger sev1 alerts. Also, GE is not displaying the permissions at all when the above is ScopesInfoDictionary is empty. Also logging every permission that doesn't have a permission info. on every request could be spiking CPU usage.